# Implement comprehensive security testing infrastructure