… __tests__/security/tests/mcp-tools-security.test.ts
… __tests__/security/tests/security-validators.test.ts
… __tests__/unit/InputValidator.test.ts
… docs/development/CONTEXT_HANDOFF_JULY_12_930AM.md
… docs/development/CRITICAL_FIXES_NEEDED.md
… docs/development/NEXT_SESSION_IMMEDIATE_TASKS.md
… docs/development/PR_225_FINAL_STATE.md
… docs/development/PR_225_NEXT_STEPS.md
… docs/development/PR_225_SECURITY_TESTING_SESSION_JULY_12_PM.md
… docs/development/QUICK_REFERENCE_PR_225.md
… docs/development/SECURITY_FIXES_APPLIED.md
… docs/development/SECURITY_FIXES_FROM_REVIEW.md
… docs/development/SECURITY_TESTING_SESSION_JULY_12.md
… docs/development/SESSION_SUMMARY_JULY_12_2025.md
… docs/development/SESSION_SUMMARY_JULY_12_EVENING.md
… docs/development/SESSION_WRAP_UP_JULY_12_EVENING.md
… docs/development/TEST_PATTERNS_REFERENCE.md
… docs/security/SECURITY_TESTING.md
… docs/security/SECURITY_TESTING_QUICK_START.md
… package.json
… scripts/security-test-runner.js
… src/security/InputValidator.ts